requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
boto3>=1.34.0
streamlit>=1.32.0
pandas>=2.0.0
